requirements.txt
arrow==1.2.2
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
binaryornot==0.4.4
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
contextlib2==21.6.0
cookiecutter==1.7.3
coverage==6.4.1
cryptography==37.0.2
Django==4.0.4
django-browser-reload==1.6.0
django-environ==0.8.1
django-tailwind==3.1.1
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==14.1.1
flake8==5.0.4
gunicorn==20.1.0
h11==0.12.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
outcome==1.2.0
packaging==21.3
phantomas==0.2.3
pip-autoremove==0.10.0
pipdeptree==2.2.1
pluggy==1.0.0
poyo==0.5.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-mock==3.8.2
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.7
requests==2.28.0
schema==0.7.5
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
uuid==1.30
whitenoise==6.2.0
wsproto==1.1.0
